fix: loki logs in workflow web (#117)
* fix(workflow-run-loki): add logic to add pipeline and site to the loki tags

workflow web wasn't able to find the logs for Work since the relevant tags were missing from the Loki logs. It also appears that the python driver may not accept logs that do not have any tags at all.

* refactor(chimefrb-workspace): add loki tags config to workspace

this is currently a empty object since I couldn't think of any useful static tags to add to the logs.